requirements.txt
groq
requests
google-generativeai
cohere
python-dotenv